Removes `init()`.
I've also decided `init()` is a bad idea. While it successfully removes a duplicated type and infers it from a value binding, it's far less efficient than `@hydrate()` and much hackier, due to the fact that the properties are initialized _after_ field initializers are run, so timing doesn't actually work the way it looks like. A decorator which is reflexively assigned is much more intuitive than an initializer which gets magically re-set at a later time.