Updates `listen()` to return the element being bound to.
This is useful for enabling or otherwise initializing the element without having to query for it separately, which is very unergonomic and kind of a pain.

This also reverts a previous commit and makes `listen()` only bind to a single element. A separate `listenAll()` method binds to multiple elements. They both assert that at least one element is found.

The return type of `listen()` is inferred from the given selector based on a parsed tag name if present. This avoids the need to cast to `HTMLButtonElement` and makes the API a little more ergonomic, if a bit magical.

I'm not sure this is an ideal abstraction given that it's very closely tied to the event binding use case. A more generic `query()` API might be more flexible and maintain similar ergonomics?