Adds state host counter example.
The purpose of this example is to demonstrate how state in a low-level component can be pushed up the render tree during hydration and maintained in a higher abstraction component, which then pushes down any updates like props with unidirectional data flow. Not super happy with how it turned out, but it's good enough I guess.

Also updates several decorators to support manual converters and improves type checking so those converters infer the parameter type based on whether it's coming from a element (via the selector) or an attribute.

I actually tried a different approach where a child component would dispatch an event with the initial count which would be captured by the parent component via a `@capture()` decorator. This likely works for components with deferred hydration, but components generally hydrate in order of the `customElements.define()` call, and since child elements will be defined before parent elements just due to import ordering, they will typically hydrate first and any event is dispatched before the parent component is ready to receive it. Since a component can't know if or when the parent component will be ready to receive the hydrated value, it needs to be available on demand. A public property seems like the best way to accomplish this and is most compatible with `@hydrate()` and `@live()`.