Adds context with example.
This is a manual implementation of a context API which provides and receives values based on elements in the DOM hierarchy. Values provided by an element can be received by any descendant elements. This provides a mechanism for passing data down the DOM hierarchy.

This is necessary for higher-order components to effectively communicate with lower-order components. Such a relationship typically looks like:

```html
<higher-order-component count="5">
  <template shadowroot="open">
    <slot></slot>
  </template>
  <div>
    <lower-order-component>
      <div>The current count is <span>-</span>.</div>
    </lower-order-component>
  </div>
</higher-order-component>
```

The `<higher-order-component />` can't easily identify or get a reference to the `<lower-order-component />` because it does not have any knowledge about what lower-order components would be interested in any provided state. It can't just look at direct light DOM children, because that's actually a `<div />` tag in this example which serves some unrelated purpose. The two solutions here are:

1. Have `<higher-order-component />` define some CSS selector with which it can identify children interested in receiving its state. For example, users could write `<lower-order-component please-give-me-the-current-count />`, which the `<higher-order-component />` could query and apply. This means every higher-order component needs to define the selectors and contract for how this state is provided.
2. Use a context API, where `<higher-order-component />` provides some context which can be received by an element in its descendants.

I think the second is much more elegant and practical. In the spririt of interoperability, I tried to implement the [community protocol for context](https://github.com/webcomponents-cg/community-protocols/blob/main/proposals/context.md). Unfortunately this doesn't really work because of hydration timing. Due to import ordering and `customElements.define()`, child components almost always hydrate before parent components (and even when they don't, we can't _assume_ that the parent component will always hydrate first). This means that when a child component hydrates, it can't assume all its parent components are providing context and there is no mechanism in that proposal to wait for parent context to become available, you just have to continuously poll for context.

Instead, I implemented my own context API. The API itself doesn't actually assume that you're using Hydrator, but it does maintain state in global maps, so it's not interoperable with other context libraries or web components. You'd need to directly depend on this context library in order to provide and receive context visible to Hydrator. It's not an ideal solution, but good enough for a proof-of-concept.