Adds "functional" API design.
This is an alternative approach to the class-based design inspired by some other frameworks. It originally came from the fact that it's weird to give users an entire class to author, but then limit most of their functionality to the `hydrate()` method. In particular, `connectedCallback()` and `disconnectedCallback()` are available, but generally shouldn't be used as its easy to shoot yourself in the foot. Also I have a personal vendetta against "initializer methods" which make classes a lot harder to work with safely. So I changed my thinking with `hydrate()` as a factory function with the class hidden from the user. The rest of the API fell out from those decisions.

I ended up using Signals, since this design required a distinct object to pass around which could reflect its changes to the DOM as I can't rely on decorators. I'm still not sold on Signals as a concept, they seem very nuanced to me and it's additional bundle size overhead which I'd like to avoid given that this library is supposed to be a very lightweight approach to making individual components interactable. I don't want a potential user to feel discouraged from using this library for even a single component because of bundle size impact. It should be fine to use this as just one part of a much more complex system.

I'm not totally convinced the functional API is a better approach. In the README there's a whole section on the implicit trade-offs between the two approaches, but there's enough nuance there as to be fundamentally incomparable. I do think the functional approach is more "interesting", so I'm leaving that as the root export and it's probably the one I'll talk about more. However, I'm still leaving the class-based approach checked-in and exported under `/class` for users to try out.

I apologize for this commit being so large, it was an experiment which got out of hand and I ended up changing a lot of stuff in the process. The class-based approach is mostly the same, however it is moved to no longer be the default option. If I could go back in time I would make each step a real commit to show my thought process, but I got a little carried away and too much of it is squashed into this one commit already for me to go back and untangle them.