Renames `Hydrator` to `HydroActive`.
I was never a big fan of `Hydrator` as a name and that package already exists on NPM. I wanted a name which evoked "hydration" and "interactivity" or "reactivity", so `HydroActive` seems like a nice blend of the two. SEO isn't ideal as I'm not the first to come up with that name, but there's no other libraries using the term so I think I should be good enough in that area.

Some other ideas include:
* HydroFlex - "Flexible" isn't really "interactive".
* HydroFlux - Cool sci-fi name but does really say "interactive" to me.
* HydroDynamic - Pretty sure I'm gonna get this question wrong on my physics final.
* AquaWiggle - 😹