WIP: Adds hooks examples and modifies the API.
This makes a number of changes:

First, `$.lifecycle()` provides easy access to run an initializer and a disposer for a single function. It doesn't really return anything because it is intended to apply side effects to the component itself.

Second, `hook()` constructs hooks for `$.use()`. This function isn't really necessary and is kind of a left over from when I was trying to automatically inject `$` in hooks so it didn't need to be passed in. I ultimately rejected that approach because it doesn't work for hooks defined after hydration (such as `$.waitContext()`) since we can't rely on a context-like global variable to hold the `$` value, as that wouldn't be attached to the async stack frame.

A number of examples are introduced to demonstrate `$.lifecycle()`, `$.effect()`, and hooks.

I'm not a huge fan of the hooks syntax, given that it's three nested functions running at distinct and precise times (hydration, connect, disconnect). Closures are very nice for variable scoping, but following which function is which can be quite complicated, especially since the hydration function is the one which actually returns the result value along with the connect function, which looks very ugly.