Adds disposed effect example.
This defines a global variable to interact with the component. It would normally lead to a memory leak (because `window.increment()` maintains a strong reference chain to `$.live()` which holds the `<span />` tag which is doubly linked to the rest of the DOM tree for the component. However `$.effect()`'s disposer is able to clean up this global variable and allow the component to be garbage collected correctly.