Adds `$.asyncEffect()`.
This ended up more complicated than I expected. I originally implemented this to provide an `AbortSignal` for users which I thought was convenient, however it turns out that signals don't work well in async contexts because after an `await`, you can't trace an accessors' invocation back to the effect that was triggered previously in the async stack.

To work around this, I allowed signals to be passed in as inputs. They are accessed and observed synchronously, and their values are passed through to the async effect. It's not elegant and requires you to "pre-declare" your dependencies, which definitely goes against signals.

I made a (perhaps misguided) attempt a making `$.asyncEffect()` throw if any signals are accessed within it (since they aren't tracked they have unintuitive effects). This is unfortunately not easily possible for the same reason that we can't track signals in an async context. We can't know that the current async execution stack previously included a `$.asyncEffect()` and should throw an error. However, I was able to get it to work by switching the default behavior. Now signals will throw by default, *unless* they are executed within an observed stack. This allows `$.asyncEffect()` to throw because nothing will start observing signal execution. The challenge with this is that it means we need to explicitly `unobserve()` any execution stack where we actually want untracked behavior. For the most part, HydroActive can `unobserve()` for you and does this for pretty much all its APIs. However, we can't do it for anything executed asynchronously, most notably a `setInterval()` callback. I imagine this would trip up new devs quite frequently, but whether or not it's better than an untracked `$.asyncEffect()` signal, I'm not sure.

Like most of my decisions so far, I'm picking the more *interesting* option of throwing by default. Time will tell if this is the right approach going forward.