Requires type for `$.hydrate()`.
This is beneficial for two reasons:
1. It requires users to be explicit about the return type, this also asserts that the element they found is actually the type they expected.
2. It requires a hard value reference to any dependency components. This effectively forces those components to be authored / imported to run this component, which forces bundlers to sort those dependencies first, making the component tree hydrate bottom up.

`$.query()` still technically works around this, but hopefully this is a good starting point?

Technically there is a contradiction with `$.hydrate('span', HTMLDivElement)`. This won't work at runtime, since there is no valid selector match which is also an `HTMLDivElement`, but the compile behavior is questionable. Is the return type `HTMLSpanElement` or `HTMLDivElement`? I think I would rather return an `HTMLDivElement` type in that situation, but for the life of me I can't get that to work with TypeScript. It feels simpler to me, so I'm probably doing something wrong, but I'll come back to it later.

Also made the same change to `$.live()`, requiring an explicit type parameter. Since you don't really live bind components there's no really an import ordering issue there, but I think it's good to keep the types consistent.