Adds templating example.
This attempts to support reusing a prerendered component multiple times in the same page, with some properties not known at render time. In this example, users can create counter components with an initial value provided in a text input. This means the counter cannot be completely prerendered or hydrated until the initial count is known.

I accomplished this by having users author a `<template data-hydroactive-tag="my-component" />` element which gets automatically cloned into the `<my-component />` element's shadow DOM if it was not prerendered with one. It then runs user hydration logic which can bind state to prerendered nodes which did not have all the necessary information. This also supports defining a "props" type for a component. These props are set as properties on the component instance, making them accessible to any generic web component system outside of HydroActive. However I also made a `factory()` function for generating a factory which requires these props to be given via TypeScript typings. This isn't perfect, as it can't do a runtime assertion that the properties are given. It also means that the component host type needs to include `| undefined` for all its properties since the component can be created without props being set. The factory function triggers hydration while the component is disconnected, which is valuable to make keep it valid when in user-space even before it is attached to the DOM.

Hydrating a disconnected component doesn't really have a good protocol yet, so I filed https://github.com/webcomponents-cg/community-protocols/issues/38 to discuss it more directly.