Adds repeated templating example.
This started out trying to make attribute hydration work, but then I realized it was easier to demo without a wrapper component and could be done by just repeating the tag in the HTML directly. I don't think this should really be done given that it delays content from being shown until JavaScript is executed and causes unnecessary layout shift. If you can render the component for each usage, you definitely should. The exception where this would be useful is if at least one instance of the component is client-side rendered.

Attributes from the template element are copied over on hydration, though any existing attributes on the element with the same name are preferred, so they can be easily overridden. I also demonstrated that light DOM "just works" as expected, and you can techncially override the shadow DOM with a custom DSD, though I don't think this is a good idea.