Adds `$.hydrateElement()` and an example of orchestrating hydration of multiple sub-components.
This is a temporary name, it should probably be named `$.hydrate()` and the existing `$.hydrate()` should be renamed to something else.

`$.hydrateElement()` queries for a given selector and then hydrates the result with the provided props and returns the component. This allows a component to trigger hydration of a sub-component at a specific time, provide the necessary props, and then receive the hydrated component. The "output" of the hydration (any returned state) is attached to properties on the web component, which can be read off the returned component or even destructured for a nicer syntax. This allows a component to force a particular sub-component to hydrate first, get the result of that hydration, and then use it to hydrate another sub-component. This gives a lot of control over hydration timing.

One nuance is that I had to remove "auto-hydration", where sub-components are no longer automatically hydrated before the parent. This makes it more likely the user will observe a dehydrated web component, but I think I can build in enough safety rails to make that less of an issue.