Adds basic testing support.
This includes a new `hydroactive/testing.js` import path which provides some simple testing support. Most notably it provides a `hydrate()` function which is very similar to `factory()`. It assigns any provided properties and then removes `defer-hydration` from the component to trigger hydration. The component's DOM can be easily asserted before and after hydration.

An interesting implementation note is that we provide a `query()` function to find a particular element, but we deliberately leave it dehydrated and note that in the type by returning an `Element` with a `__hydrated__` tag. This is useful to test the application's render logic and what is visible to the user before hydration has happened. Then, when the user calls `hydrate()` we assert it to be an instance of the `__hydrated__` tag meaning we're able to protect the user at compile time from misusing a dehydrated element.

One philosophical challenge with HydroActive in particular is that testing very tightly couples rendering logic to the JS component. You can't execute a HydroActive component unless it has been prerendered correctly. This means the test needs to serve prerendered HTML, you can't just render the component client-side from scratch. This is further complicated by the fact that HydroActive doesn't have knowledge of your rendering pipeline, so we can't really solve that the problem for you. Even the most basic unit test requires prerendering the component (which is dependent on your rendering stack) followed by executing the JS test. Given that HydroActive doesn't know anything about the rendering pipeline, we can't really provide tooling to make that easier for you, though `rules_prerender` sure could ;).

For this repository, I set up `@web/test-runner` with a custom plugin to serve hard-coded HTML files which include test cases. This uses Mocha + Chai to handle most of the test infrastructure, with Sinon for clock mocking. TBH, I would rather just have Jasmine or something as an all-in-one test framework compatible with this browser environment, and maybe I'll circle back to that shortly. Since the test HTML is hard-coded it does duplicate logic with the example HTML files, meaning these aren't the most complete tests. I think this is fine for now given that the examples are intended to be super simple, however it does indicate that any non-trivial project would likely need to figure out their pipeline of "render a component" -> "test the rendered component" in order to have any confidence in their rendering logic.

Aside from the infrastructure, I only added tests for a couple components just to prove it could work. I'll add more tests as we go.