Configures Jasmine for use in `@web/test-runner`.
This sets up Jasmine to run tests and report the results back to the runner. Unfortunately it got a bit complicated. There is [a plugin](https://github.com/blueprintui/web-test-runner-jasmine/) to do this, but we can't just use it out of the box because it uses `testFrameworkHtml` under the hood, which is compatible with my devserver plugin approach as they both attempt to specify the page HTML, when only one can. So instead, I had to copy a lot of the implementation and integrate it into `serveHtmlPlugin` (now named `jasminePlugin`).