Adds `hit()` and refactors `hydrate()` and `query()`.
The `hit()` function is a wrapper of Mocha's `it()` function which queries for the test case in the DOM and then cleans it up afterwards. This provides better test isolation because components are removed when they are no longer being tested and any associated resources are cleaned up.

This commit also moves the class reference from the `query()` (which is now hidden in `hit()`) to the `hydrate()` function. This greatly simplifies the flow the type system and defers the class reference until the point where it is needed ("needed" as in, the class needs to be defined, not as in we need to actually do anything with the class). This allows us to remove `query()` from public API and makes `hydrate()` more intuitive. This also removes all the `Dehydrated` typings which are no longer needed.

Normally, I'm not a fan of having optional or unneeded dependencies for an NPM package, and it's usually a sign to me that multiple packages should be used. Strictly speaking, `hit()` depends on `it()` which is a Mocha function, therefore IMHO we should have `@hydroactive/core` for everything we currently have, and a separate `@hydroactive/mocha` for the `hit()` function. However, I'm bending the rule here for a couple reasons:
1. I (probably foolishly) published the initial package as `hydroactive`, so we don't currently have an NPM scope. Definitely solvable, I just don't want to deal with it right now.
2. `hit()` is just a single function, not a huge amount of code to really justify a different package.
3. We don't _actually_ have a dependency on `mocha` because `it()` is typically exposed globally, so all we need to do is type check which only requires a dev dependency.

If the testing module expands in the future, then I think it would definitely be worth a separate package.