Replace `factory()` with `hydrate()`.
After playing around some more, I've realized that `factory()` is just a user-space hydration function. It finds a template to clone its own DOM and then immediately hydrates. However it is unnecessary to clone a template and then hydrate, when it really doesn't matter where the component comes from (whether cloned from a template, or just prerendered directly into the page). As a result, I decided to push out the template into user-space, and focus on a `hydrate()` primitive, which takes any element with its class and props, and hydrates it. This is effectively the testing utility I made previously, but used more generically.

This primitive allows users to render an inactive component to a template, clone that template, hydrate the result with required props, and then append it to the DOM. I think this is a lot more flexible than the existing templating examples and makes them much more intuitive.

Unfortunately there is one very awkward edge case which comes from the fact that template contents are held in a different "inert" `Document`, and so are their clones preventing custom element upgrades. I had to handle this by adopting the element into the current document, upgrading it, and then adopting it back into its original document. This seems to be the "least suprising" approach compared to alternatives, though it is definitely awkward.