Defines custom elements directly in `component()`.
By defining the custom element before returning the component class to user-space and by requiring the component class for any hydration operations, we are effectively forcing developers to define HydroActive classes prior to hydration. This avoids the foot gun where a user attempts to hydrate a component before its definition has loaded and silently fails, breaking the next use of that component which expects it to be fully hydrated.

This should effectively prevent such a case for HydroActive components, however we cannot guarantee this is always the case because users could hydrate non-HydroActive components which do expose their class reference independently of the `customElements.define()` statement. Such a situation _should_ be rare, given that any component which is defined in the top level scope adjacent to the class definition will work out naturally.

Also made two minors changes:
1. The tag name is transformed to UpperCamelCase and used to override the class name. This makes it more clear what class a developer is looking at when debugging, and also makes error messages more obvious. Unfortunately, this name comes from the tag name, not the variable defined by the user. For example, if the user writes `const Foo = component('bar-baz', () => { /* ... */ });` the component will be named `BarBaz`, _not_ `Foo`. This could be counter-intuitive, but I think it's better than nothing.
2. The class definition is out of the `component()` function as much as possible. 99% of the class is unrelated to the `component()` function, it only needs a reference to the input `hydrate()` function. This is done by making the class abstract and subclassing it for each component, only override the minimum necessary to close over `hydrate()`. This makes things a little more maintainable but doesn't really affect behavior.