Renames `$.hydrate()` to `$.read()`.
The purpose of this function is to query the DOM for an element, read some state off it, and convert that state to a useful representation (parse to a number for example). This is independent of hydration, which is handled by `$.hydrateElement()` and will be renamed to `$.hydrate()`.

I'm not super-happy with `$.read()` as a name, but it's the best I can come up with. Alternatives considered include:
* `$.init()` - Implies the target is being initialized or modified in some way.
* `$.parse()` - Implies text is being processed, which isn't accurate.
* `$.convert()` - Verbage is weird.
* `$.from()` - Too vague about what "from" means in this context.