Updates `$.query()` and `$.queryAll()` to throw when they find a custom element.
Custom elements are not hydrated by `$.query()` and could leak dehydrated elements into user-space. Users should call `$.hydrate()` instead which will handle this more safely.

The selector is not sufficient to identify custom elements because the tag name may not always be present. `$.query('#my-button)` is _probably_ a `button`, but could just as easily be a `material-button` and skip hydration. The error message suggests an alternative `$.hydrate()` call instead.