Removes type check on `$.query()` selector and fixes `$.read()` of custom elements.
The type check on `$.query()`'s selector was always a bit hacky and it's not that useful given that we now have a runtime assertion which is more comprehensive. In the process, I also noticed that `$.read()` was actually needed to get references to eagerly hydrated sub-components. We can't `$.hydrate()` an eagerly loaded sub-component because it has already hydrated, so `$.read()` now returns custom elements if found, but will assert that they are already hydrated (by checking the `defer-hydration` attribute and checking `instanceof`). Only deferred sub-components need `$.hydrate()`.