Allows binding booleans to attributes.
Whenever the signal boolean changes, a bound attribute is updated to either empty string or removed from the element altogether. This has the effect of specifying whether or not an attribute is set at all on an element based on the boolean. This is very useful for boolean attributes like `hidden`.