Adds `$.scope`.
This creates a new "scope" for subsequent `query`, `live`, or any other utilities to use when looking up selectors. This allows a component to temporarily scope `querySelector` operations to a subtree of its component. This is useful when hydrating from multiple DOM structures with a similar shape, such as multiple rows of a table.