Deletes the core source code.
This was always intended to be a prototype, not production code. It lacks any meaningful tests, includes two different syntaxes, 95% of the implementation lives in a single source file, and is generally a mess. The codebase has outlived its original purpose and I'd like to do a clean rewrite which incrementally adds tests and features more deliberately than the prototype.