Initial signals implementation.
This is a larger commit for sure, but it builds out an initial implementation of signals and effects. This is a bit different implementation than the original approach, using more objects and less functions. It defines a `Producer` / `Consumer` graph where `signal` creates a `Producer` to provide its value and `effect` creates a `Consumer` to consume those values. This should be scalable for computed values and other use cases.

I also added a new concept of a `Scheduler`. This allows `effect` to be decoupled from the actual timing of the action it invokes. For now there is only a `MacrotaskScheduler` which uses `setTimeout` as well as a `TestScheduler` for testing purposes. In the future, I would expect to see a `UiScheduler` using a batched `requestAnimationFrame` call and maybe even a `MicrotaskScheduler` using `queueMicrotask`. For now, this should be good enough though. Requiring an explicit `Scheduler` makes `effect` a little more annoying to work with, but most users likely won't call it directly, but instead go through `ComponentRef` or another wrapper which binds the `effect` lifecycle anyways, so this shouldn't be too big a problem.

I tested manually and confirmed that signals and effects do correctly clean themselves up and are eligible for garbage collection. This will be an important invariant to maintain going forward which is difficult to test automatically.