Adds `cached` primitive.
This provides a cached value computed from other signals.

There are two ways this could be implemented:
1.  When an upstream `Producer` changes, mark this `cached` value dirty as well, and notify downstream `Consumer` values.
2.  When an upstream `Producer` changes, immediately rerun this `cached` and check if the result changed. Cache it, and only notify downstream `Consumer` values if it result changed.

The first seems closer to what a function named "cached" should do. It's also the behavior of Angular's `computed`, so that's what I went with here. I suspect the second might actually be a different primitive I'm struggling to name right now.

I considered adding `equals` but it actually doesn't make sense at all with this kind of signal primitive. A `cached` is marked dirty before it is executed, so it has no opportunity to check an equality function before notifying consumers that it changed (doing so would be the second design).