Adds `UiScheduler`.
This is useful for scheduling operations likely to mutate the DOM. The scheduler uses `requestAnimationFrame` to wait for the opportune moment before a render to adjust the DOM.

I considered batching multiple scheduled actions into a single `requestAnimationFrame` call, but decided not to do so as it complicated the implementation more than I was comfortable with and I'm not sure the value added justifies the effort. Batching is something which can be considered in the future, but likely not a major issue initially.