Adds custom equality function.
This allows signals to define a custom function to determine whether or not they are dirty when modified. If they are not dirty, downstream consumers are not notified as an optimization, since nothing has meaningfully changed.

Defaults to `Object.is` so the behavior is equivalent to `===` when no custom function is provided.