Adds `optional` parameter to `ElementRef.prototype.query{,All}`.
This allows these functions to default to assuming that the selector will find at least one element, and if not, to throw an error. This makes typing easier because we don't have to assert that an element was found in the general case.

When you do have situations where an element might not exist, `{ optional: false }` configures `query{,All}` to allow this case. `query` required some function overloading to configure the return type correctly.

I'm not 100% sold on `optional` for `queryAll`. I think it is reasonably intuitive for `queryAll` to return an empty array, and doing so doesn't complicate the typings. I decided to move forward with this anyways just to try it. However based on feedback we may want to reevaluate whether it makes sense for `queryAll` to default `optional` to `true`, or to just drop the parameter altogether.