Adds `ComponentRef.prototype.bind`.
This allows "binding" a reactive signal to a specific element in the DOM, automatically updating the element's text content whenever the signal changes. The typing got a bit involved here, but I think it's all working now and has some extensive typing tests to ensure we do not regress in that area.