Adds `ComponentRef.prototype.bindAttr`.
This is roughly equivalent to `ComponentRef.prototype.bind`, except applied to a specific attribute, rather than text content. This is 95% the same as `bind`, so most of the implementation is shared. However it is slightly complicated by the need to check if the attribute is already bound and then write to the attribute, both of these necessitate the shared implementation does a little bit less work than it really wants to.