Adds `ElementSerializer`.
This adds support for a distinct `ElementSerializer` type which serializes to/from a complete element, not just a specific attribute or text content. This allows for more complex serialized types, such as a `<my-user id="1234" name="Devel">` element, where multiple attributes need to be read and written to at once.

This is very strongly typed, so an `ElementSerializer` is not only specific to a particular `Value` type, but also the `Element` subtype it requires. The active element is also inferred from the provided query in cases like `comp.live` and `comp.bind`.

The types are starting to get very complicated and I'm hesitant to add too much more to them. Fortunately we have a sizeable amount of type regression tests, which should hopefully catch anything which goes wrong in the future. This actually required a load-bearing variance annotation `ElementSerializer<Value, in El extends Element>`. I don't fully understand exactly why that's necessary, but it is needed for `comp.live` to compare serializer types correctly to the queried element type.