Adds `nextFrame` testing utility.
`waitForAnimationFrame` was becoming too useful and annoying to copy-paste, so this seems like a good candidate for a shared utility.

This also fixes a slight bug where `waitForAnimationFrame` would run before RAF callbacks scheduled after it, even if they are in the same animation frame. `nextFrame` now handles this correctly and resolves after all RAF callbacks in its batch have been invoked.

Unfortunately using `setTimeout` breaks Jasmine's mock clock, which requires an explicit tick. I think the best we can do is automatically tick in `nextFrame`. The dependency on Jasmine is loose, so this can become a public API without necessarily coupling it to Jasmine.