Refactors `parseHtml` to be a little more strict with custom elements.
Now it requires a value reference of any parsed custom element classes to be passed in as an input. This ensures no undefined components are included and inactive at runtime, which can be annoying to debug. The return type is also inferred from a provided class.

This is a little unergonomic for non-custom elements use cases such as `parseHtml(HTMLDivElement, '<div></div>')`, however this should be relatively rare in user code, since they will almost always be parsing a rendered component.

One potential shortcoming is that `parseHtml` is not composable, so if one component requires another to be rendered, `dependencies` cannot be easily reused between them in a useful way. This might be a future problem to tackle or potentially considered out-of-scope given that `parseHtml` is intended as a utility for simple testing use cases, not a full rendering engine. A problem for another day.