Adds `Scheduler.prototype.stable`.
This converts the `Scheduler` interface into an abstract class which includes functionality for `isStable` and `stable`. The implementations of `Scheduler` subclasses are (relatively) unaffected by this complexity.

The most interesting part here is that it is surprisingly difficult to await until the scheduler is stable because it is difficult to ensure that a new action was not added to the scheduler between the final action executing and the `stable` `Promise` resolving. Added some extra tests to be especially sure that this behavior is right.