Adds `HydroActiveComponent.prototype.stable`.
This allows any user of a HydroActive component to wait for a component to be "stable", essentially waiting for all pending effects to be invoked.

I'm not totally convinced this is a good idea. I think it's kind of bad practice for production code to rely on this, but it is definitely necessary for testing and I can see it being occasionally necessary in production. I think my main concern is that we have a few competing interests:
1.  It should be possible to batch multiple signal set operations into a single effect execution for improved performance. - Means we can't execute effects synchronously on set.
2.  Test timing should align with production timing. - This is difficult because `requestAnimationFrame` is (relatively) slow and we don't want a `flush` function which devs can forget to invoke. Also would necessiate synchronous effects if we _don't_ have some kind of `flush` or `await stable()`.
3.  Tests should not be dependent on specific component timing. - Switching a component from `comp.live` to `comp.bind` or something else should not break all its tests querying the DOM.
4.  A component's stability should also guarantee on its children are stable. - Avoids the problem of Lit's `await el.updateComplete` where every child element needs to be individually awaited.
5.  A HydroActive component should not "leak" its internals, including that it was implemented with HydroActive. - This is especially difficult with respect to timing of DOM execution, how `stable` is exposed, and how it interacts with child elements.

I don't have answers for all of these yet, but I suspect there is a fundamentally intractable problem here. We'll just need to find the best trade-off we can in this complicated space. For now, I'm adding a simple `stable` method, but we'll see how this evolves in the future.