Adds `hydrate`, `comp.hydrate`, and `comp.hydrateAll` functions.
These functions make it easy to hydrate specific elements as dependencies of a component. To discuss a few notable design decisions:

`comp.hydrate` queries an element, meaning it's not possible to separate the query step from the hydration step. I don't like this, but for now it seems to be necessary. The obvious alternative would be `comp.query('some-comp').hydrate(SomeComp)`. That looks nice, but it begs the question: What does `comp.query('some-comp')` return? I don't want it to return `ElementRef<SomeComp>`, because the element is not yet hydrated and does not actually satisfy the contract of `SomeComp`. `ElementRef<Dehydrated<SomeComp>>` is more agreeable, but relies on typing for safety, meaning it doesn't help JavaScript users. We could allow `ElementRef<SomeComp>` but then throw on any `read` operation, but I don't like the idea of having an `ElementRef` instance you can't read, that feels like a design smell for a type that doesn't provide any abstraction. Instead my goal is to prevent `ElementRef<SomeDehydratedElement>` as a possible pattern, but this requires `comp.hydrate` to do the query step.

`hydrate` and `comp.hydrate` both require the class definition of the element (`SomeElement`) as an input. This doesn't actually do anything, it only does an `instanceof` check to make sure the right class was given. However the point is to ensure that the custom element is properly defined prior to hydration. If we didn't do this, the `customElements.define` call might be moved *after* hydration and fail at runtime. This ensures bundlers always order the `customElements.define` call prior to any associated `hydrate` call. We're using `class SomeElement {}` as a proxy for `customElements.define('some-element', SomeElement);`, but it is the closest we can do here, and usually accurate as custom elements are typically defined in the top-level scope of the file containing their class definition.

`hydrate` unfortunately does not throw an error if a component's hydration fails. This is due to the fact that `attributeChangedCallback` does not propagate its errors to whomever invoked `el.removeAttribute('defer-hydration')`. This is just the way that custom element reaction callbacks work. If I could have it propagate errors I would, but this doesn't seem possible right now without encoding special behavior into the `defer-hydration` community protocol. I started [a discussion(https://github.com/webcomponents-cg/community-protocols/pull/15#issuecomment-1962284394)] on this topic, but until a relationship is formalized, I don't think there is a way to support propagating hydration errors.